requirements.txt
docx
google-generativeai
tkinter